b=11171 r=jitendra,cliffw
Some documentation (s/.pl//g) fixes.
Add a $rslt_loc variable to set the location of the $rslt file.  Defaults
to /tmp for backward compatibility.
Cleanup some syntax/style:
- put $ next to variable names in awk scriptlets
- one space on either side of |
Make sure the amount of data the sgp_dd is going to write actually fits in
the device and suggest reducing $size and exiting with an error if it does
not.