b=11171 i=cliffw,jitendra
Create a more portable way of remote shelling:
- set $DSH variable to either rsh or ssh (default)
The location of the results can be set with rslt_loc which defaults to /tmp.
Eliminate the "scp brutal hack" needed for non-shared filesystems.
A related work item could be to do a test to see if $rslt_loc is actually
shared between the nodes and not use remote_shell if it is indeed shared.
Specifically invoke gawk in obdfilter-survey where it uses gnu awk extensions