JBD: Fix NULL pointer bh->b_data on NUMA box with journal checksumming.
Current journal checksumming patch failed fsstress test on NUMA. The
bh->b_data passed to the crc32_be () function could be NULL pointer,
which caused kernel oops immediately when running fsstress with -o
journal_checksum. It is because the page is part of highmem on NUMA box.
We need to kmap the page before access the bh->b_data to calculate
the checksums.