b=14286 i=grev i=adilger branch=HEAD
change the rundbench script to use contents of DBENCH_LIB variable to
find the location of the dbench executable (like sanity.sh and
sanity_quota.sh do) so that we use the same environment variables to specify the location of the dbench executable in all the test scripts.
instead of using different environment variables to configure the
same thing in different scripts.