b=10600 r=adilger r=johann
Branch b1_6
qunit size will be changed when quota limitation is too low/high;
record the pending quota write in order to get more accureate
quota; delete the patch for bug12588, which is unnecessary when
this patch is landed. This bug also contains 14526, 14299, 14601
and 13794, which are found and landed during v1_4_12.