b=15482 i=cliffw i=wangyb
Allow a kernel target to specify a delimiter between the final "flavor"
specification and the version.
Specifically this allows RH kernels to be <version>smp and SLES10 kernels
to be <version>-smp.
Maintaining the vendors versioning specification means that when software
is trying to determine a vendors kernel using patterns (like OFED does), it
properly detects the SLES10 kernel.
This all lays the ground work for bug 15316.