b=15870 i=adilger
Backing this out as it tries to solve a problem that doesn't really exist
and causes build breakage in doing so.

I have advised the customer that opened this bug as to how they can fix
thier OFED source pool to match the interface provided by the
kernel-ib-devel RPM, which is our reference implementation.