Branch b_head_libcfs b=16420
Adds a new library for userspace, multithread utilities. It inlcudes nidstrings.c
and parser.c (taken from lustre/utils).

The new include (libcfsutil.h) also includes platform.h which used to be in lustre/utils.
This includes defines for locks and condition variables, and eventually should be consolidated with
libcfs.h when that supports multithreaded apps.