b=17025 r=adilger,nikita
If a client creates and opens some files, and then mds gets crashed, there exists a chance that the objids in OSTs will be used for multiple files.

The root cause of this issue is that we don't have the appropriate mechanism to handle the case if the replaying of open request fails.

This is a hot fix for this problem. By skipping the orphan objects, the client won't see that multiple files using the same objid any more.