b=16312 i=robert.read i=yibin.wang
Make lbuild build kernels based on the vendor's own SRPM rather than all of
the gyrations we used to go through to build the kernel tarballs and so on.

This should give people a kernel (and associated packages) that looks much
more like the kernel (and packages) they are used to getting from their
vendor.

This has resulted in a much easier to understand lbuild and once we finally
fully drop support for rhel4 and sles9 we can get rid of the
lustre-kernel-2.4.spec.in, build/lbuild.old_school and build/lmake files.
Anyone know knows anything about the build system will understand what kind
of simplicity in understanding of how the build works that will yield.