b=17671 i=zhen.liang i=sheng.yang
Enable Lustre to be built with OFED 1.4.1.

This is just the glue needed for the build with OFED 1.4.1 to actually work.
OFED 1.4.1 won't actually get built with a Lustre branch until a target file
specifies:

OFED_VERSION=1.4.1-rc4

Note that OFED 1.4.1 only supports RHEL4, RHEL5 and SLES10 targets.  SLES9
users are out of luck (but have been for some time with OFED).