b=19163 i=wangyb i=rread
Because we rename our repackaging of the vendor kernel to include "lustre"
in the name, this changes the packages implicit Provides: attribute.  So we
make our repackaged RPMs provide the %name-%version-%release explictly so
that our kernel RPMs will satisfy the vendor's other RPMs' Requires:
attributes.

Introduce the "$BUILD_GEN" variable.  This allows the build software
engineers to force the reuse hash to be a new unique value when they update
the build software in a way that will change the resulting kernel RPMs.