b=17605 i=rread i=grev
Use the mpicc wrapper for building MPI programs rather than trying to
discover all the bits and pieces of the installed MPI library.

Due to the way the autotools work, this requires putting the sources
into their own subdir, so that the CC command can be overridden.

Fix the pointers to in-tree MPI programs in test-framework.sh.