b=19975 i=yangsheng i=wangyb
Because we now build RHEL5 and SLES10&11 kernels from their own spec files,
we don't produce a kernel-source for RHEL5.  RH's story to people who want
source is to unpack and patch the kernel SRPM.  So we should provide the
SRPM for people who want kernel source.

SLES just comes along for the ride (actually, as do RHEL4 and SLES9 but we
don't support them for server on 1.8, so no need to distribute any kernel
[S]RPMs for them whatsoever.