b=19721 i=adilger
Add a debian/ dir to allow building on Debian/Ubuntu systems.  The process
is basically (from the top-level) lustre dir:
$ dpkg-buildpackage
$ sudo m-a build ../lustre-source_1.8.1-1_all.deb

The above needs to be captured in a "make debs" Makefile target so that it
works like the "make rpms" target, including building lustre packages as
well as a binary kernel modules package.  This will be the next step in this
enhancement.

All of this has really only been tested to build the patchless client with
Ubuntu's 2.6.24-19-generic kernel at this point.  However the bits necessary
to build a patched server kernel are included, even if they need to be tested
and perhaps tweaked for supported kernels.  At least theoretically, it should
work for newer kernels even.