Branch HEAD b=21240,19242
Disabling replay-single 44a, which is using obd_race, which appears to be racy itself, and
also appear to be incorrectly checking the return code of lclt recover, which appears be supposed
to fail in this test. Also disabling 44b which looks like it has the same issue with lctl recover.