b=21514 i=wangyb i=yangsheng
If lbuild is interrupted (i.e. LBATS "Cancel") then it should not call
the ccache -s exit trap action because experience has shown that the
ccache binary is gone by the time lbuild gets to calling it.

Also, extend the API for the delete_exit_trap() to handle multiple
action handles in a single call.  This makes it easy to create a stack
of actions and then delete them all in a single command.

What would be even nicer would be (shell and pdsh style) glob handling
so that one could do delete_exit_trap foo[1-10], assuming one pushed 10
actions on a stack to be unwound during an operation that is supposed to
be atomic.  Or more simply delete_exit_trap foo*.