Introduce .gitignore files.
The top level .gitignore file is new and is an attempt at
pulling in some of the common items that you might get for
free (by default)with CVS.  The other subdir/.gitignore files
are translated versions of their corresponding .cvsignore
file.  Because CVS does not descend into a subdir when
applying a ingore rule we have to prepend a "/" to
the pathname to get git to behave the same way.