b=21519 fix race between ldlm_pools_thread and ldlm_space_free
After ldlm_space_free dropped the refcount of a namespace to 0, it
may start to cleanup data structures in the namespace, while the
ldlm_pools_thread can still pick up the freeing namespace. This patch
is to instruct ldlm_pools_thread to skip namespace which is being
freed.

i=oleg
i=wangdi