b=21829 fix broken llobdstat and add a counter parameter
Need to make sure we limit the search for OBD stats files to the obdfilter
subdirectory of "/proc/fs/lustre".
Add a counter argument to limit the number of items returned when using the
interval parameter.
Fix lots of whitespace atrocities as well as better format some of the code.

i=cliff
i=wangyb