b=21490 1.8<->2.0 interop: in 2.0 server and 1.8 client env. read 1. 8 file hung
The fix for bug 18631 introduced an interop bug for 1.8.x clients with 2.0 OSTs.
The problem was that a zero-length RMF_RCS was included in the OST_BRW replies
for READs, which caused the RPC header to have eight more bytes for storing the
length of that zero-length buffer.  The fix is to split RQF_OST_BRW into two,
one for READs and one for WRITEs, with RMF_RCS present only in the reply for
OST_BRW_WRITE.

i=robert.read@sun.com
i=tom.wang@sun.com