b=21924 create test dir before starting changelogs
i=rread
A test-level directory (d0.lustre-rsync-test) was recently added to the $tdir path.
This directory is created before the tests are run.  But lustre-rsync requires the
source and target systems to be identical before the changelog is started; the
presence of this directory in the source, missing in the target, causes the
failure seen here.  This patch simply creates $tdir on the source and target
before each test is run, insuring a coherent starting state.