b=19336 Get rid of the EXTRA_VERSION_DELIMETER shenanigans
We used to carry around a bunch of baggage in order to specify what kind
of delimeter to put between the version and "extra version".  The truth of
the matter is that this should always be "-".

This patch includes some support for a build system developer to force
an uncached rebuild of all products.

i=wangyb
i=yangsheng