b=16909 Use INFO/WARN instead of WARN/ERROR for the slow messages.
We should use INFO/WARN instead of WARN/ERROR for the slow messages.
Not only is there no real error here but it fixes an annoying quirk
of the message formatting.  With the old levels you would see the
messages formatted differently based on the time.

  Lustre: lc1-OST0001: slow parent lock 289s due to heavy IO load
  LustreError: 0-0: lc1-OST0001: slow parent lock 324s due to heavy IO load

With the new levels things are more consistent.

  Lustre: lc1-OST0001: slow parent lock 289s due to heavy IO load
  Lustre: lc1-OST0001: slow parent lock 324s due to heavy IO load