b=20953 a fix for test_30 of sanity-quota
lfs calculates time left to the end of a grace period based
on current client's time and MDS-based grace period end which
can lead to cetain anomalies in lfs output when the client's
clock and the MDS' clock are not synchronized. This patch
introduces some additional I/O so that grace time decisions
and calculations will only be made on the MDS side.

i=Elena Gryaznova