b=21452 support for weak-modules
Add support to our RPM SPEC for the weak-modules script.

This requires that we install our modules under
/lib/modules/$(uname -r)/updates/kernel.  I think this is the correct
location for us in any case given that we are a kernel "addon" package.
Relax the kernel Requires: to work better with weak-modules.
Use the external dependency generator as the internal one cannot deal
with kernel modules.

i=mjmac
i=wangyb