b=20619 be flexible in version format
The version format for the OFED release needs to be more flexible than
the previous M.m format that was required.  Now it can be any dot-separated
format.  i.e. 1 or 1.4 or 1.5.1 or 1.5.1.2, etc.  Of course, it has to be a
valid and unique OFED version (so the "1" example would be a valid version
string but invalid because there is no OFED version "1".  Ditto for 1.5.1.2
at the time of writing this comment.).

i=michael.macdonald
i=minh.diep