b=20997 Enable time limit on the request to be queued
i=rread,nathan

This feature is enabled by setting the time limit on the request. The limit
is on how long the request is queued, whether or not it's ever been sent.
IOW, it's not a "resend" limit, because it can expire even if the request
has never been sent the first time.

Patch removes also unused imp flags and modify conf-sanity test to use 'mds
mount first' scheme to avoid long test time