b=22481 Add "lfs_migrate" script from manual into lustre/scripts and RPMs
Add enhanced lfs_migrate script from manual into lustre/scripts directory
and in RPM packages.  It does a "poor man's" migration of files from
their current OST layout to a new OST layout as chosen by the MDS.
It is currently not safe to use for files that are being written to.
It will result in the current file becoming an open-unlinked file, and
the migrated file will be used for any subsequent opens.  Any writes to
either the old or migrated file will in clients seeing different data,
and writes to the old file will be lost when it is closed.

There is no direct control of object allocation within the script, it
depends on external control to manage space.  Use "lfs find" to
select appropriate files to migrate from full OSTs. "lctl --deactivate"
is needed on the MDS to deactivate OSTs if trying to empty an OST
completely, otherwise QOS will tend to migrate files to empty OSTs.

i=brian.murrell
i=nathan.rutman