b=20355 formalize support for pthreads
BG/P needs proper pthreads library arguments for linking, so properly
support pthreads properly in configure, using AX_PTHREAD.
The need for the pthread subdirs is because you cannot adjust the CC
variable value used on a per target basis, only for an entire Makefile.
This is the same problem we had with using the MPICC for some targets.
So to make some targets compile with the CC that the AX_PTHREAD macro
discovers/sets, we need to use a new Makefile.  File locations remain
the same however.

i=panda (b1_8 version)
i=whitebear (b1_8 version)