b=21452 kABI tracking
Changes required to build packages that properly interact with
the kABI tracking included in SLES and RHEL kernels:
- take local copies of the find_requires generators and make modif-
  ications to them so that they can find the kernel artifacts installed
  in our build tree
  + thusly, override the __find_requires macro in all of the RPM
    build operations

Fix package versioning of patchless packages to also include the
timestamp, just like patched packages do, for non-release builds.
Also, remove the _lustre from the kernel version in the lustre
package names for patchless packages.

Note that changes needed to enable the external dependency checker
and thus include kABI dependencies in the RPM are not included here
for RHEL5.  Once RH have updated their kABI whitelist to include the
symbols we need, the patches in bz attachments 30546-7 need to land
for RHEL5 weak modules to work and a bug needs to be filed with OFA
to land the change in bz attachment 30547.

i=wangyb
i=mjmac