b=23185 check for both arches
When we build our version of the SLES kernel, we optimize it for i686
whereas the SUSE kernel is i386.  The actual arch makes a difference in
where the Module.symvers can be found, so just look in both locations
to cover both the upstream vendor kernel as well as our patched kernel.