b=22612 dig out Module.symvers for sles10
When we build modules in lbuild for sles10 kernels, we point to the
contents of the kernel-source RPM.  That does not contain the
Module.symvers file though and thus we get a nice big warning that
the modules built (i.e. ldiskfs) will not contain versions.
This is because we need to unpack the /usr/src/linux-*obj/Module.symvers
file from the kernel-$flavour or kernel-syms RPM into the location pointed to
by --with-linux-obj.

Fix a bug in unpack_linux_devel_rpm-$DISTRO where we were not checking
for the success of the function.

i=mjmac
i=wangyb