b=22967 s/LB_LINUX_CONFTEST/AC_LANG_CONFTEST/
There was a macro update in more recent releases of autoconf that
requires us to use AC_LANG_CONFTEST instead of the LB_LINUX_CONFTEST
that we currently use.  The problem using LB_LINUX_CONFTEST causes
is that as configure is determining capabilities and setting capability
flags, these are not being used when compiling further conftest.c
programs.

So for example if a macro determines if foo is available and then sets
FOO if it is, and then a test in a subsequent macro tries to use FOO,
it will find it undefined.

i=michael.macdonald
i=andreas.dilger