b=22418 2.6.32 compat synchronize_rcu already defined
In 2.6.32 kernels when !CONFIG_TREE_PREEMPT_RCU synchronize_rcu is a #define
instead of being exported symbol.  In 2.6.33 kernels it  is an exported as
a symbol regardless of the configure option.  To handle this when the symbol
is not exported use the kernels provided define if available.  Otherwise
fallback to providing our own for older kernels.

o=brian.behlendorf
i=andreas
i=kalpak