b=23701 remove a variable from stack, allocate from slab
Through the assembly of mds_open from the crash dump, it seems,
although the iattr var declaration is put into a compound statement,
the stack for the variable is reserved from the beginning to the
end of mds_open processing. According to crash, the size of iattr is 80 bytes.

i=Oleg Drokin
i=Alexander Zarochentsev