b=20744 use Novell supplied ofed-devel
Several changes to support using ofed-devel supplied in SLES10 SP3
(currently as PTFs).
Fix a bug with passing arguments to rpm2cpio that should not be passed.
Add a hook for distro specific module building after the kernel[-devel]
build and unpack.
Fix resolve_arch to be able to resolve either the "canonical" arch for
i?86, or the arch that the patchless kernel is expected to be.
Bail out of download_ofed earlier if using "inkernel".
Fix a bug where an rpm2cpio command was not being error checked.
Fix a bug where we were not processing and making available files
in all of the RPMS/i?86 subdirs.

i=mjmac
i=wangyb