b=24089 avoid reuse cache storage collisions
This is a hack.  Basically, all nodes do their build and the last one
to store in the cache wins.
A better solution would be for nodes to announce their intention to
build for a given hash and any other nodes that will do the same wait
for the first node to finish and store it.  Subsequent nodes then just
use the first node's stored result.

i=wangyb
i=cliff