b=14836 Fail if "lfs setstripe" has positional parameters
Since 1.6.1 it has been possible to use named options -c/-i/-s to
specify the layout parameters for "lfs setstripe".

Since 1.8.0 the use of the old positional parameters has generated
a warning that they are deprecated.

The code was set to fail building at 2.1.0.0, so this needs to be
resolved before moving to a release build.  Instead of simply removing
the conditional code, change "lfs setstripe" to return an error if
positional parameters are used.  That would avoid the confusion of
creating files with names of the old {size}, {offset}, and {count}
parameters, instead of returning a meaningful error to the user.