b=20563 Fix fid_flatten32() to not lose OID bits
The original implementation of fid_flatten32() was broken due to an
error in the shift calculation (note to self - "0x00" is 8 bits, not
16 bits).  This could negatively impact 32-bit clients that were
creating more than 64k files in the same directory.

This 32-bit inode number is visible only within a single client mount,
is not used in any persistent storage, and only if a 2.x server is in
use (which is basically none today) by a 32-bit client, so there is no
issue to change it at this time.