b=24037 avoid conflicting DCACHE_LUSTRE_INVALID
In newer kernels our definition of DCACHE_LUSTRE_INVALID needs
to be changed to avoid conflictint with other DCACHE flags.
Use a high value to avoid conflict for a long time, maybe until
we can actually get rid of the need for it entirely.

i=sheng.yang
i=vitaly.fertman