LU-103 Add support for test groups.
Added an -g GROUP_FILE option to auster. The tests listed in the
group file will be used instead of any passed on the command line. I also added
a couple of test-groups to get things started, although
I'm sure we'll want to refine these.

Took care of a few cleanups in run_suite as well.

Signed-off-by: Robert Read <rread@whamcloud.com>
Change-Id: I5000c2b95a6eb1a49f5305637d4b55de5edc6d5c
Reviewed-on: http://review.whamcloud.com/278
Tested-by: Hudson
Reviewed-by: Chris Gearing <chris@whamcloud.com>
Reviewed-by: Wei Liu <sarah@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>