LU-210 backout debian packaging with patches
The debian packaging process included the [de-]applicaiton of patches
that were the commits made since the last tag.  This provided greater
transparency as to exactly what was included in a release when it was
made between tags.

Unfortunately the process that was being used to achieve this failed
in a scenario where there were patches to files which were later removed
from the tarball making process.

For what it's worth, the right way to correct this and keep the
transparency is to have "make dist" create the tarball from the most
recent tag made and then have the patches generated that bring the
source up to current HEAD.  In order to achieve this however both debian
and RPM packaging would need to operate in this manner.  That would
mean adding Patch$n: lines to the RPM spec for each of the generated
patches.  I'm not really sure we want to go there though.

Signed-off-by: Brian J. Murrell <brian@whamcloud.com>
Change-Id: I9242799f3987e68e806e4c398e06cecbe1f5cc27
Reviewed-on: http://review.whamcloud.com/425
Tested-by: Hudson
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Reviewed-by: Michael MacDonald <mjmac@whamcloud.com>