LU-326 make sure df send statfs request
in test_18c of recovery_small.sh, it uses df to send a statfs request
to OST to detect the disconnection stat, but it will fail to send
the request if there is a cache of statfs in client node.
the patch adds a "sleep 1" to make sure the cache of statfs expires
before df sends statfs request.

Change-Id: Ia1f56b8160e77827d5249656488734d7defd6c81
Signed-off-by: Hongchao Zhang <hongchao.zhang@whamcloud.com>
Reviewed-on: http://review.whamcloud.com/551
Tested-by: Hudson
Reviewed-by: Johann Lombardi <johann@whamcloud.com>
Reviewed-by: Fan Yong <yong.fan@whamcloud.com>