LU-64 define LOV default layout as FID_SEQ value
For directories without any default striping information the MDS
sends a magic LOV_OBJECT_GROUP_DEFAULT value for the lmm_seq field.
Since this is really a magic FID_SEQ value, define it in the same
place as other FID_SEQ values as FID_SEQ_LOV_DEFAULT so that it is
visible.

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I0ae9dc993d15aedb38ab79e40c67a687d956a6c4
Reviewed-on: http://review.whamcloud.com/910
Reviewed-by: Prakash Surya <surya1@llnl.gov>
Tested-by: Hudson
Tested-by: Maloo <whamcloud.maloo@gmail.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>