LU-384 Add space to default PATHS
The update in c81b47a724dd4414b8bd79eb99b9737a70fd5855 broke
configure if --with-linux was not specified, causing build
failures because the .config file could not be found.  Fix this.

Update some ancient documentation that is referenced if the
.config file cannot be found to better describe what is needed
to build Lustre with modern kernel packages (for RHEL at least).

Change-Id: I5b373e3ee3ef42c68257c27a4b194c9e00cf023f
Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Signed-off-by: Brian J. Murrell <brian@whamcloud.com>
Reviewed-on: http://review.whamcloud.com/946
Tested-by: Hudson
Reviewed-by: Johann Lombardi <johann@whamcloud.com>