LU-540 lloop: fix LLOOP_INFO to return FID
Instead of returning an inode number to userspace for the LLOOP_INFO
ioctl, instead return a FID value.

This does not pose compatibility problems, because the ioctl() number
was already different between Lustre 1.x and 2.x, and the lctl needed
for using lloop devices is always built with the kernel modules.

Change-Id: Ia3ee6dab830ad65e2a11d9248982c82ce76221f1
Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-on: http://review.whamcloud.com/1149
Tested-by: Hudson
Tested-by: Maloo <whamcloud.maloo@gmail.com>
Reviewed-by: Jinshan Xiong <jay@whamcloud.com>
Reviewed-by: Fan Yong <yong.fan@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>