LU-429 Fix and quiet debug message in filter_connect
The comment refers to mds_connect(), which no longer exits.  So I
just removed the comment.

The message says "Received MDS connection", but this function is
used when clients connect to OSTs as well.  We don't need to see
this message on the console all of the time, so I changed it to
D_INFO.

Change-Id: Iebfbaafe39df702862ade126b979528855173d5c
Signed-off-by: Christopher J. Morrone <morrone2@llnl.gov>
Signed-off-by: Yang Sheng <ys@whamcloud.com>
Reviewed-on: http://review.whamcloud.com/963
Tested-by: Hudson
Tested-by: Maloo <whamcloud.maloo@gmail.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>