LUDOC-11 cleanup: cleanups of examples and text
Clean up some examples seen in the manual while working on LUDOC-14.

When the --index parameter is passed to mkfs.lustre, it will already
assign an OST or MDT index, so OSTffff will not be in the output.

Clean up the example for multiple NIDs to remove the use of the old
elan network type.

Signed-off-by: Richard Henwood <richard.henwood@intel.com>
Change-Id: Ibe1068d66627e72a88c2584f413d2c40df99cab0
Reviewed-on: http://review.whamcloud.com/2181
Tested-by: Hudson
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>