LUDOC-11 Misc cleanups of examples and text
Clean up some examples seen in the manual while working on LUDOC-14.

Use "ost0" and "ost1" in examples, to match the actual OST indices
used, instead of "ost1" and "ost2".  Same for "oss0" and "oss1", etc.

Use "lctl {get,set}_param" instead of direct /proc access in a few
places (there are a lot more that should be fixed).

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: Ibeb4e3318b6fbdb8a81283d5cfad14a69e27443b