LU-762 ldiskfs: don't drop directory nlink to 0
When landing the nlink patch for ext4, for an unknown reason the
logic in ext4_dec_count() was changed from the ext3 version of
the patch.  It now drops the nlink = 0 temporarily and then
if it is a directory with nlink == 0 it increases nlink again.

Instead, only drop nlink if it is larger than 2.

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: Id4dd16d3764efdf6a7ea29da92de51398b300370
Reviewed-on: http://review.whamcloud.com/1579
Reviewed-by: Johann Lombardi <johann@whamcloud.com>
Tested-by: Hudson