LU-601 mdd: cleanup error messages and code style
While looking at LU-601, I added comment blocks for several of
the functions in the PENDING directory cleanup path, along with
fixes to the code style and error messages.  No functional changes
should be introduced by this code change.

The error message that was formerly printed by the 1.8 MDS in
mds_unlink_orphan() and checked by replay-single.sh test_37 no
longer exists in the 2.x MDS code.  Re-add a message if there
is an error cleaning up directory orphans, and change test_37
to match the resulting error message (similar to the 1.8 code).

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: Ic97843e49643041513f3b2fc6a40f2be63ffddca
Reviewed-on: http://review.whamcloud.com/1679
Tested-by: Hudson
Tested-by: Maloo <whamcloud.maloo@gmail.com>
Reviewed-by: Fan Yong <yong.fan@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>