LU-322 tests: Pool test fixes for large number of OSTS
This patch fixes issues with large numbers of OSTs with the
ost-pools test. we need to use hexidecimal numbers for
the pool args since the OSTs UUID are named with hex numbers.
Currently the test does pass but erros can be seen in the logs
when more than 9 OSTs exist.

Signed-off-by: Minh Diep <mdiep@whamcloud.com>
Change-Id: I5db7540ee710c528af33d768a105a55bd09229b3
Reviewed-on: http://review.whamcloud.com/1841
Tested-by: Hudson
Tested-by: Maloo <whamcloud.maloo@gmail.com>
Reviewed-by: Yu Jian <yujian@whamcloud.com>
Reviewed-by: James Simmons <uja.ornl@gmail.com>
Reviewed-by: Johann Lombardi <johann@whamcloud.com>