LUDOC-11 upgrade: document 2.1 upgrade requirement
The 1.8->2.x upgrade section was referring only to Lustre 2.0, but
should refer to Lustre 2.x more generically.

Update the minimum 1.8 version for 2.1 interoperability to 1.8.6.

Remove statement that allows both 1.8 and 2.x servers to be running
at the same time, since this is not tested.  Instead, require that
both the MDS and OSS nodes are upgraded to 2.x at the same time.

Remove references to ancient 1.4.x Lustre versions, and distinctions
that are only present for versions of Lustre 1.6.5 and older.

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I94a41610f65e597accc773f534e518f4e87fcab0