LU-1148 build: Handle download kernel-devel failure
When the kernel-devel fails to get downloaded with yumdownloader,
emit a meaningful error message rather than just mysteriously exiting.
While it would be nice to actually be able to do something about this
failure programatically, typically it's something that requires
human intervention somewhere.  Or at least more time than we'd rather
have a build blocked for while it's rectified.

Signed-off-by: Brian J. Murrell <brian@whamcloud.com>
Change-Id: I070968f73eb0e78152705e78eedefdc3b6ac8e4f
Reviewed-on: http://review.whamcloud.com/2226
Reviewed-by: Minh Diep <mdiep@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Tested-by: Hudson
Tested-by: Maloo <whamcloud.maloo@gmail.com>