LU-1221 ldiskfs: Replace sysname with nodename in MMP
sysname holds "Linux" by default, i.e. what appears when
doing a "uname -s"; nodename should be used to print the
machine's hostname, i.e. what is returned when doing a
"uname -n" or "hostname", and what
gethostname(2)/sethostname(2) manipulate, in order to
notify the administrator of the node which is contending
to mount the filesystem.

Andreas says this was introduced when porting the MMP
patches from RHEL5 to RHEL6, and then also pushed upstream
to ext4; a patch for upstream ext4 has already been
submitted.

MRP-222
Signed-off-by: Nikitas Angelinas <nikitas_angelinas@xyratex.com>
Signed-off-by: Niu Yawei <niu@whamcloud.com>
Change-Id: Ie5c8e99ae536c2239134140589d678d3a4f92639
Reviewed-on: http://review.whamcloud.com/2337
Tested-by: Maloo <whamcloud.maloo@gmail.com>
Reviewed-by: Johann Lombardi <johann@whamcloud.com>