LUDOC-2 add "--raw" option to "lfs getstripe"
A new option to "lfs getstripe" is introduced; the "--raw" or
"-R" option.

If this option is specified, the stripe information is printed
without substituting the filesystem's default values for
unspecified fields. If the striping EA is not set, 0, 0, and -1
will be printed for the stripe count, size, and offset respectively.

Signed-off-by: Yu Jian <yujian@whamcloud.com>
Change-Id: I2d4a1bf8fd063afb40345609db9c8204d6e3cc09