LU-56 libcfs: implementation of cpu partition
This patch is implementation of CPU partition (CPT) and CPT table.

CPU partition (CPT) is conceptually similar to the 'cpuset' of Linux.
It is designed to provide an abstract layer that can be conveniently
used by kernel threads, unlike the 'cpuset' that is most accessible
from userspace.

CPT table is a set of CPTs:
- A CPT table can contain 1-N CPU partitions
- CPUs in one CPT shouldn’t overlap with CPUs in CPTs of the same
CPT table
- A CPT table can cover all CPUs in the system, or a subset of system
CPUs.

Signed-off-by: Liang Zhen <liang@whamcloud.com>
Change-Id: Idd3168f21ea82891d11dffe06cbb386e1505889c
Reviewed-on: http://review.whamcloud.com/2461
Tested-by: Hudson
Reviewed-by: Doug Oucharek <doug@whamcloud.com>
Tested-by: Maloo <whamcloud.maloo@gmail.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>