LU-1481 build: quiet warnings due to := assignment
When running automake to generate Makefiles, newer compilers will
warn about non-portable assignments:

  lvfs/autoMakefile.am:52: `:='-style assignments are not portable

The use of ":=" vs. "=" determines when conditional expressions
are evaluated, but have absolutely no use for constant assignments,
so there is no point in using them at all in that case.

Convert all ":=" to "=" for constant assignments, leaving the
decision to use ":=" for conditional assignments until later.

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I76caedbc1368a0d2fe02137c46b46c36fc3a500c
Reviewed-on: http://review.whamcloud.com/3129
Tested-by: Hudson
Reviewed-by: Brian J. Murrell <brian@whamcloud.com>
Tested-by: Maloo <whamcloud.maloo@gmail.com>
Reviewed-by: Brian Behlendorf <behlendorf1@llnl.gov>
Reviewed-by: Oleg Drokin <green@whamcloud.com>