LU-549 llite: Improve statfs performance if selinux is disabled
Even if selinux is disabled, client still tries to get selinux
attributes from MDS. As xattrs are not yet cached, this significantly
slows down xattr heavy operations like ls -l. This patch forces
to return -EOPNOTSUPP on the client side if selinux is disabled.
It speeds up ls -l 25% for cold-cache case and 50 %for hot-cache
case.

Signed-off-by: James Simmons <uja.ornl@gmail.com>
Signed-off-by: Yevheniy Demchenko <zheka@uvt.cz>
Change-Id: I9c52ce820b4d3c3d518673cb35e1c23acefb3d8d
Reviewed-on: http://review.whamcloud.com/3257
Tested-by: Hudson
Tested-by: Maloo <whamcloud.maloo@gmail.com>
Reviewed-by: Fan Yong <yong.fan@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>