LU-1329 ptlrpc: resend request on -EINPROGRESS
It seems the EINPROGRESS is going to be used for many different
purposes (e.g. on statfs see review 3198). As a result, it sounds
like ptlrpc is the right place to resend requests on EINPROGRESS.
Upper layers can still decide to handle EINPROGRESS by themselves by
setting rq_no_retry_einprogress to 1.

Signed-off-by: Johann Lombardi <johann@whamcloud.com>
Change-Id: Iae2a5976666e66be4f6e71f82c5653e5636ba07d
Reviewed-on: http://review.whamcloud.com/3262
Tested-by: Hudson
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Tested-by: Maloo <whamcloud.maloo@gmail.com>
Reviewed-by: Niu Yawei <niu@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>