LU-1608 utils: require --fsname for mkfs.lustre
When formatting a filesystem, require that fsname is specified to
mkfs.lustre so that users have to think about this a bit.  In some
cases, two filesystems at the same site have used the same default
filesystem name, and this later caused problems when trying to mount
both filesystems on the same node.

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I15a5ebb8d7d3afc6037587961e94dc66426bcf5d
Reviewed-on: http://review.whamcloud.com/3297
Tested-by: Hudson
Reviewed-by: Yu Jian <yujian@whamcloud.com>
Reviewed-by: Li Wei <liwei@whamcloud.com>
Tested-by: Maloo <whamcloud.maloo@gmail.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>