LU-1842 protocol: add support for OBD_IDX_READ
This patch defines a new RPC format (namely OBD_IDX_READ) which
allows to read the content of an index file via a bulk transfer.
It is simlilar to MDS_READPAGE except that it is not tied to a
specific key/record format (readdir relies on lu_dirent/lu_dirpage).
Like readdir, key/record pairs are stored in a container of a
fixed size (i.e. 4KB) regardless of the client & server page size.

Signed-off-by: Johann Lombardi <johann@whamcloud.com>
Change-Id: I34071ca05a3bd4e6c01bfe4fc533ab79bdd98b3c
Reviewed-on: http://review.whamcloud.com/3942
Tested-by: Hudson
Reviewed-by: Fan Yong <yong.fan@whamcloud.com>
Tested-by: Maloo <whamcloud.maloo@gmail.com>
Reviewed-by: Niu Yawei <niu@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>