LU-1842 osp: osp-on-ost device
Setup osp-on-ost device for each OST to maintain the connection
to MDT0, quota slave & ofd will send quota or fid sequence request
over this connection.

We also setup one such device on MDT0 for now, it'll be used by
the quota slave on MDT0, once the DNE code landed, this device
(on MDT0) could be replaced by normal osp device.

Signed-off-by: Wang Di <di.wang@whamcloud.com>
Signed-off-by: Niu Yawei <niu@whamcloud.com>
Change-Id: I2229cec1fa206dcce65a4cceae032566f4c20901
Reviewed-on: http://review.whamcloud.com/4068
Reviewed-by: Johann Lombardi <johann.lombardi@intel.com>
Tested-by: Hudson
Tested-by: Maloo <whamcloud.maloo@gmail.com>