LU-1842 quota: add quota locks support on QMT
This patch adds methods for the quota lock lvb management and
implements an intent policy for quota locks.
It also introduces the rebalance thread which is in charge of
sending glimpse callbacks to slave to notify them of new quota
settings.

This patch also enables sanity-quota.sh test 36 which should now
pass.

Signed-off-by: Johann Lombardi <johann.lombardi@intel.com>
Change-Id: I0cf7f86d480296f7e8fe774ed8937aae27736f88
Reviewed-on: http://review.whamcloud.com/4166
Tested-by: Hudson
Tested-by: Maloo <whamcloud.maloo@gmail.com>