LU-1943 tests: Refresh conf-sanity 32[ab]
Existing conf-sanity 32[ab] does not run on multi-node clusters or
network types other than TCP.  This patch rewrites the tests to
start the targets on the MDS and mount the upgraded file system on
the primary client.  This scheme works both on a single-node
development environment and a typical Autotest cluster.

The disk image tarball format has been updated to include more
"metadata".  For example, the 1.8.7-wc1 and 2.1.1 tarballs added
by this patch include the kernel versions, architectures, Lustre
versions, and commit SHAs (currently only Lustre version) used to
create the disk images.  In addition, the script used to create
the tarballs is recorded as a test in test_32newtarball().

A couple of new data verifications are added to make sure the name
space, the file data, and most of the file attributes are consistent
across upgrades.

Signed-off-by: Li Wei <liwei@whamcloud.com>
Signed-off-by: Alex Zhuravlev <alexey.zhuravlev@intel.com>
Change-Id: I3b0d38d67e86e0e2ac24d9aae4406b055b63ce61
Reviewed-on: http://review.whamcloud.com/4213
Tested-by: Hudson
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Tested-by: Maloo <whamcloud.maloo@gmail.com>