LU-2111 echo_cleint: Wrong error code returns for MDS echo
If MD echo client fails to resolve a path a error is
returned by PTR_ERR but the rc variable is never updated
with this value to return back thus the application
thinks it worked. This patch properly sets the returned
rc variable to let the application testing know a failure
has occured.

Signed-off-by: James Simmons <uja.ornl@gmail.com>
Change-Id: If7d2e9fbb28bcb239f7cc5021efebdcf0784ea14
Reviewed-on: http://review.whamcloud.com/4225
Tested-by: Hudson
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Tested-by: Maloo <whamcloud.maloo@gmail.com>
Reviewed-by: Liang Zhen <liang@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>