LU-2012 tests: disable replay-dual test_14b
Disable replay-dual.sh test_14b until OST gap handling is fixed.

The test is modified to make the pass/fail result more clear for when
it is re-enabled, since any blocks that are allocated on the OST while
the test is running will cause it to fail.  This could happen because
of updated config llog, OI updates, etc.  Allow some small margin of
space to be allocated on the OST before declaring failure.  To make
failed the orphan handling totally clear, ensure the orphan object is
much larger than the margin of error.

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: Ib14ecdfa1f8fbd5807195acb60e5ba507f500c1e
Reviewed-on: http://review.whamcloud.com/4237
Tested-by: Hudson
Tested-by: Maloo <whamcloud.maloo@gmail.com>
Reviewed-by: Mike Pershin <tappro@whamcloud.com>
Reviewed-by: Yu Jian <yujian@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>