LUDOC-66 lfsck: improve online LFSCK documentation
Improve the description of the online LFSCK feature, providing a
high-level overview of the feature of what the LFSCK feature is,
and improving the readability of the text.

Also update the Backup/Restore section to replace bugzilla reference
with Jira bug reference.  Also mention that the OST device label needs
to be reset for 2.4 and later filesystems.

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: Iadbdec4f17ca7bc1ad49065756088e4d05ddcab0
Reviewed-on: http://review.whamcloud.com/4506
Tested-by: Hudson
Reviewed-by: Fan Yong <yong.fan@whamcloud.com>
Reviewed-by: Richard Henwood <richard.henwood@intel.com>