LU-1064 mds: fix mds_lookup lma removal error path
In commit 1fd243c89e3b221d40ce74b8ef47f1bca760c8f9 if an error is hit
removing the "lma" xattr from an updated 2.x inode, then the open
transaction handle would never be committed, and the MDS would hang.

This is unlikely to be a problem, as the only errors fsfilt_set_md()
will hit that are not programming bugs are due to IO errors from the
underlying disk (which is an even bigger problem).

Make sure that the transaction is committed, even after an error.

Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: If7d19a2337a12efafacd20d5c4e5c00e85300c1e
Reviewed-on: http://review.whamcloud.com/4729
Tested-by: Hudson
Reviewed-by: Iurii Golovach <Iurii_Golovach@xyratex.com>
Tested-by: Maloo <whamcloud.maloo@gmail.com>
Reviewed-by: Johann Lombardi <johann.lombardi@intel.com>