LU-2498 mount: Set 'deadline' scheduler on ldiskfs
The recommended IO scheduler to use with Lustre on top of ldiskfs is
either the 'deadline' scheduler, or the 'noop' scheduler. This patch
attempts to automatically configure a device to use the 'deadline'
scheduler for the user at mount time. This makes it unnecessary to use
any out of band mechanism to ensure this value is set properly (i.e.
start up scripts, kernel configuration, etc.).

Also, if the scheduler is already set to 'noop', this patch will leave
that setting in place. This makes it easy for a user to manually set the
scheduler to 'noop', if they would prefer to use that scheduler instead
of 'deadline'.

Signed-off-by: Prakash Surya <surya1@llnl.gov>
Change-Id: If869468ba5370afddb29233dd893d74b9c45f367
Reviewed-on: http://review.whamcloud.com/4853
Tested-by: Hudson
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>
Tested-by: Maloo <whamcloud.maloo@gmail.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>