LUDOC-104: Design elements to highlight lustre specific features.
The Docbook 5 standard allows for a condition attribute to all
elements. By setting this condition with the value of l23 or l24
passages of the manual that are specific to version of Lustre
can be marked up.

This change adds an annotation against elements that have the
condition attribute set. Graphics are rendered in-line at the
relevant point in the text and against the relevant entries in
the table of contents.

'condition' annotations to two existing lustre specific sections
of the to demonstrate this new design feature.

The document design elements are displayed similarly across epub, html,
xhtml and pdf renderings of the manual.

Signed-off-by: Richard Henwood <richard.henwood@intel.com>
Change-Id: Idd2cb762e219204bfc63c9ed81ccedd49b2c43ac
Reviewed-on: http://review.whamcloud.com/4611
Tested-by: Hudson