LU-1484 lprocfs: handle hidden proc_dir_entry users
The RHEL 5.9 2.6.18-348.1.1.el5 kernel uses both the new style
"pde_users" and the old style "deleted" flags for marking a
proc_dir_entry as deleted.  Unfortunately, the new "pde_users"
data is hidden in an external structure that is not visible to
the callers or in the headers (for binary compatibility I guess?)
so our configure checks cannot find it.

Instead, just check for proc_fops == NULL in a racy manner on
such kernels, since we cannot do locking and the locking is
mostly just needed as a memory barrier since pde_fops could
become NULL at any time after dropping the lock.

Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Change-Id: I05cae305e24ffff09a06e3ad17c28c175c3ebbe5
Reviewed-on: http://review.whamcloud.com/5253
Tested-by: Hudson
Tested-by: Maloo <whamcloud.maloo@gmail.com>
Reviewed-by: Johann Lombardi <johann.lombardi@intel.com>