LU-3059 lu: shrink lu_object by 8 bytes on x86_64
Remove the lo_depth member from struct lu_object.  This field is never
set and only read in lu_object_print().  Remove the lo_flags member.
This field was only used in lu_object_alloc() and can be replaced with
an on-stack mask to keep trace of which layers have been allocated.

Signed-off-by: John L. Hammond <john.hammond@intel.com>
Change-Id: I998cfda0234f4c418227c9ba640850a78ca7893e
Reviewed-on: http://review.whamcloud.com/5890
Tested-by: Hudson
Tested-by: Maloo <whamcloud.maloo@gmail.com>
Reviewed-by: Alex Zhuravlev <alexey.zhuravlev@intel.com>
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>