LU-3041 lfsck: improve lfsck.sh on b1_8
The test directory of lfsck.sh contains some files referencing same
object, which could cause error when removing the directory on test
cleanup. Also, in lfsck.sh we shouldn't use debugfs to remove objects,
that'll cause quota usage inconsistence warning in e2fsck.

This patch is a backport of commit 622148 and 2df010, and it also
includes part of the patch of LU-3133 and LU-3180.

Test-Parameters: testlist=lfsck

Signed-off-by: Liu Ying <emoly.liu@intel.com>
Change-Id: I6fd5db0c921e744f12a92676eb906d730b9c3d10
Reviewed-on: http://review.whamcloud.com/5968
Reviewed-by: Niu Yawei <yawei.niu@intel.com>
Tested-by: Hudson
Tested-by: Maloo <whamcloud.maloo@gmail.com>
Reviewed-by: Jian Yu <jian.yu@intel.com>
Reviewed-by: Johann Lombardi <johann.lombardi@intel.com>