LUDOC-146 benchmark: water-down obdfilter-survey warning.
The obdfilter-survey uses a different object sequence (2) compared to
normal filesystem objects (currently always 0), so the two do not
collide.

The manual now indicates that obdfilter-survey may not destroy
production data. A brief description of the object sequence difference
between obdfilter_survey and normal objects is included.

The task to provide evidence to reduce the vagueness of this warning
is LU-3444.

Signed-off-by: Richard Henwood <richard.henwood@intel.com>
Change-Id: Iac5fd3d9a89906dd04290ab62b2cff43ece29e53
Reviewed-on: http://review.whamcloud.com/6416
Tested-by: Hudson