LU-3576 mdc: prevent fall through in mdc_iocontrol()
In mdc_iocontrol() add a goto to the end of the LL_IOC_HSM_STATE_SET
case, preventing fall through into the next case. In the same
function, replace the return statement in OBD_IOC_QUOTACTL with a
goto, so that a reference to the module is not leaked.

Lustre-commit: 0d166ef1fd80b0056497969b956d04e32e86aa85
Lustre-change: http://review.whamcloud.com/6962

Signed-off-by: John L. Hammond <john.hammond@intel.com>
Signed-off-by: Bob Glossman <bob.glossman@intel.com>
Change-Id: I762de2aaf794f1691cb243feed5b65289f91e1dd
Reviewed-on: http://review.whamcloud.com/7276
Tested-by: Hudson
Tested-by: Maloo <whamcloud.maloo@gmail.com>
Reviewed-by: Aurelien Degremont <aurelien.degremont@cea.fr>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>