LU-3648 test: large_xattr_enable doesn't detect large xattrs
The large_xattr_enable function looks for "large_xattr" from
the output of dumpe2fs to determine if large xattrs is
enabled. When large extended attributes are enabled, the file
system flag "ea_inode" is set and not "large_xattr".

Since "large_xattr" may be used in the future for this
attribute, the fix is to grep for both "ea_inode" and
"large_xattr".

Signed-off-by: James Nunez <james.a.nunez@intel.com>
Change-Id: Iebe744e7fc70c878f1897b853e37f85cd6e59031
Reviewed-on: http://review.whamcloud.com/7133
Tested-by: Hudson
Reviewed-by: Keith Mannthey <keith.mannthey@intel.com>
Reviewed-by: Jian Yu <jian.yu@intel.com>
Tested-by: Maloo <whamcloud.maloo@gmail.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>