LU-3633 test: Sanity test 101d fails due to full OST
Sanity test 101d checks the available space on the file
system against the size of the file before writing, but
the file stripe is never checked or set. So, test 101d
has failed with enough space on the file system to write
the file, but the default stripe was 1 and there was not
enough space on that single OST for the file.

This patch sets the stripe of the file to -1 (all OSTs)
before writing to the file system.

Signed-off-by: James Nunez <james.a.nunez@intel.com>
Change-Id: I69b8f24bc6e6fe05c7179160bf89610516b9c5cb
Reviewed-on: http://review.whamcloud.com/7179
Reviewed-by: Niu Yawei <yawei.niu@intel.com>
Reviewed-by: Keith Mannthey <keith.mannthey@intel.com>
Tested-by: Hudson
Tested-by: Maloo <whamcloud.maloo@gmail.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>