LU-2757 dynlocks: move into osd-ldiskfs
Commit 43603bd1 (LU-1548 osd: move i_htree_lock to iam container)
eliminated the use of dynlocks inside ldiskfs. Now that it's only
used in the osd-ldiskfs code, we can move it under osd-ldiskfs/,
and eliminate a patch we apply during build. Long term dynlocks
will never merged into the ext4 work so it makes sense to move it
to the osd-ldisk layer. Also the mdd layer was carrying techinical
debt of dynlocks as well so this patch removes it from that layer.
This work helps us move forward to a patchless ldiskfs in the
future.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>
Signed-off-by: James Simmons <uja.ornl@gmail.com>
Change-Id: I3d2ca2317426df3a7f3855d85a834a37091a78bd
Reviewed-on: http://review.whamcloud.com/5282
Reviewed-by: Bob Glossman <bob.glossman@intel.com>
Tested-by: Hudson
Tested-by: Maloo <whamcloud.maloo@gmail.com>
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>