LU-2158 lvfs: remove the fsfilt and lvfs layer
The fsfilt/lvfs layer has been obsoleted by the osd
layer. This patch moves the remaining fsfilt code
related to ldiskfs to osd-ldiskfs where it belongs.
The remaining bits of the lvfs code it moved to the
ptlrpc layer where it is used. The last of the procfs
code is integrated into the obdclass layer as well.

Signed-off-by: James Simmons <uja.ornl@gmail.com>
Signed-off-by: Mikhail Pershin <mike.pershin@intel.com>
Change-Id: Ie72103639fbcbbcedd0f810ac68044b38a0f0b65
Reviewed-on: http://review.whamcloud.com/5512
Tested-by: Hudson
Tested-by: Maloo <whamcloud.maloo@gmail.com>
Reviewed-by: Jian Yu <jian.yu@intel.com>
Reviewed-by: Bob Glossman <bob.glossman@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>