LU-4056 hsm: reorder MDT proc cleanup
In mdt_procfs_init() remove the call to ptlrpc_lprocfs_register_obd().
The single file that this provides (../mdt/*/stats) is backed by the
MDT OBD's obd_svc_stats which are not tallied. In mdt_procfs_fini()
remove the corresponding call to ptlrpc_lprocfs_unregister_obd() and
move the call to hsm_cdt_procfs_fini() to before
lprocfs_obd_cleanup(). This prevents a use after free of the entries
in the MDT's hsm subdirectory.

Signed-off-by: John L. Hammond <john.hammond@intel.com>
Change-Id: Iad6e7738047377660f0077c9368f79d96f903d02
Reviewed-on: http://review.whamcloud.com/7846
Tested-by: Hudson
Tested-by: Maloo <whamcloud.maloo@gmail.com>
Reviewed-by: Faccini Bruno <bruno.faccini@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>