LU-2456 lnet: DLC user space Configuration utility
This is the eighth patch of a set of patches that enables DLC.

This patch adds an lnet configuration utility, lnetctl,
which provides a way for sys admins to configure LNET
parameters.  This utility has been added to the /lnet/utils
directory

The utility is built only if libyaml is installed on the
system.

For SLES libyaml is not available, so if DLC is to be
built libyaml has to be installed manually.

Signed-off-by: Amir Shehata <amir.shehata@intel.com>
Change-Id: I81e8eef3452af0e7a246ca78213dd1a2d4b59775
Reviewed-on: http://review.whamcloud.com/8026
Tested-by: Jenkins
Reviewed-by: James Simmons <uja.ornl@gmail.com>
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Nathaniel Clark <nathaniel.l.clark@intel.com>
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>