LUDOC-114 lnet: LNET/LND credits and flow control
Summary: Even though the bug is about credits and flow control,
this commit is more about giving a better understanding of
lustre networking and lnet routing. This main topic has also
been touched upon in two files. Big changes to two files and
a small addition to LustreTuning.xml. I have tried to improve
on the  definition of what a lustre network is and how the
phrase-lnet is used differently in different contexts. If time
had permitted, I would have made an illustration for lnet
routing, but I couldn't. It would help to introduce one there
and one in the UnderstandingLustreNetworking.xml file to
illustrate the point that lustre networks are different from
real networks.

Signed-off-by: Avadhani J <avadhanijonnavithula@gmail.com>
Signed-off-by: Linda Bebernes <linda.bebernes@intel.com>
Signed-off-by: Richard Henwood <richard.henwood@intel.com>
Change-Id: I7592c5b35f61102050112e1b0e6ac4e042371b2b
Reviewed-on: http://review.whamcloud.com/7555
Tested-by: Hudson
Reviewed-by: Amir Shehata <amir.shehata@intel.com>