LU-3985 tests: Archive access must be done from an agent
In sanity-hsm, all access to the archive backend must be
done from an agent. Tests 111a and 111b call the
copytool_remove_backend routine that calls find on the
archive backend. With this patch, the call to find is
done on the agent with a call to "do_facet $SINGLEAGT".

Signed-off-by: James Nunez <james.a.nunez@intel.com>
Change-Id: I537fe5e019861908612140b3c96ea2018a738e46
Reviewed-on: http://review.whamcloud.com/8076
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: John L. Hammond <john.hammond@intel.com>
Reviewed-by: Jinshan Xiong <jinshan.xiong@intel.com>
Reviewed-by: Aurelien Degremont <aurelien.degremont@cea.fr>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>