LU-4277 lod: handle os_state as a flag, check READONLY
Handle the osfs->os_state values as flags instead of as values.
Otherwise, if an OSD has OS_STATE_DEGRADED and another flag set
it will not match to skip the OST during object allocation.

Similarly, there was no check for the OS_STATE_READONLY flag in
the create path, which should be used to always skip an OST for
object creation until it is cleared.

These flags are not yet set by osd-zfs, only by osd-ldiskfs so far.

Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Change-Id: I0ef59a9fd631a6942d963e37cfce2bfe183ebbe5
Reviewed-on: http://review.whamcloud.com/8378
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Alex Zhuravlev <alexey.zhuravlev@intel.com>
Reviewed-by: Fan Yong <fan.yong@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>