LU-3974 client: use standard dentry locking function
Over time the struct dentry reference count have change
its data structure. In the 3.11 kernel it was changed
to a struct lockref. To handle these cases in the past
Lustre implemented a d_refcount wrapper to handle all
these changes. At the same time for 3.11 a d_count()
wrapper was introduced. So this patch moved use to use
this d_count wrapper and if not present we emulate it.

Signed-off-by: James Simmons <uja.ornl@gmail.com>
Change-Id: I4c175e789298841f9d599e5195dbd94ccea6216a
Reviewed-on: http://review.whamcloud.com/7741
Reviewed-by: Peng Tao <bergwolf@gmail.com>
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Bob Glossman <bob.glossman@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>