LU-4307 kerberos: fix __exit usage and unknown symbol errors
For linux all functions marked with __exit should not be
called as a cleanup function in the module init functions.
This patch removes __exit from the gss module exit functions
that are being used as cleanup in the module init function.
The second issue addressed is the unknown symbols due to
gss_[init/exit]_keyring being called when HAVE_GSS_KEYRING
is not defined.

Signed-off-by: James Simmons <uja.ornl@gmail.com>
Change-Id: I2de7a3ade47f585f5ec1634ee55c1d2ce56f2921
Reviewed-on: http://review.whamcloud.com/8481
Tested-by: Jenkins
Reviewed-by: Nathaniel Clark <nathaniel.l.clark@intel.com>
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Thomas Stibor <thomas@stibor.net>
Reviewed-by: Andrew Korty <ajk@iu.edu>