LU-4343 tests: mkdir failing in sanity-hsm test 228
sanity-hsm test 228 calls mkdir on $tdir. Currently, the tdir
variable is two directories. This is changed in LU-2524. Until
LU-2524 lands, any call to mkdir with the tdir variable needs
the "-p" flag.

Also added removal of two files that the test creates and a new
routine to create small files with dd using the sync flag.

Signed-off-by: James Nunez <james.a.nunez@intel.com>
Change-Id: Ic78a14625d8c45f405c013b96c84da909b9b9244
Reviewed-on: http://review.whamcloud.com/8542
Reviewed-by: John L. Hammond <john.hammond@intel.com>
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Faccini Bruno <bruno.faccini@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>