LUDOC-210 jobstats: Clarify jobstats section and add ALPS
Clarified the jobstats section of the documentation by separating the
environment variable values for jobid_var from the special values for
jobid_var.  The special values are now explained in paragraphs with
examples following the table.  The table's columns were swapped
because it is really the job scheduler which determines the value of
jobid_var.  The description preceding the table was clarified.  A
run-on sentence was split up at the beginning of section 12.2.1, and
the article "a" was added before client.

Signed-off-by: Ryan Haasken <haasken@cray.com>
Change-Id: I9f92891ad44ab3e86d81042d0978d32d9603ecfd
Reviewed-on: http://review.whamcloud.com/8577
Tested-by: Jenkins
Reviewed-by: Patrick Farrell <paf@cray.com>
Reviewed-by: Chris Horn <hornc@cray.com>
Reviewed-by: Richard Henwood <richard.henwood@intel.com>