LU-3467 ofd: remove obsoleted OBD methods
Remove unused OBD methods from OFD, keep those needed
for echo client and rename them to indicate that

ofd_get_info() and ofd_set_info_async() are only needed
for local calls with single key.

Clean the usage of fti_fid, don't copy FID to it when
it is possible to point on FID in request buffers directly.

Signed-off-by: Mikhail Pershin <mike.pershin@intel.com>
Change-Id: I5622e14bccacb809bca1c10499c23bcaf72e2a68
Reviewed-on: http://review.whamcloud.com/8580
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Alex Zhuravlev <alexey.zhuravlev@intel.com>
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>
Reviewed-by: Fan Yong <fan.yong@intel.com>