LU-4217 build: bump build warnings to 2.7 development
The "acl" mount option on the client has been deprecated since
Lustre 1.8 (using ACLs is enforced by the admin on the MDS).
Unfortunately, there are still configs using this option, so
it cannot be removed yet, or it would cause an error at mount time.

The DNE cross-MDT locking depends on cross-MDT rename support, so
is also pushed to 2.6.53, but is expected to be fixed sooner.

Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Change-Id: Idcfe68eaec12df08538f1479a13c2c208c3ebbe5
Reviewed-on: http://review.whamcloud.com/8627
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>