LU-4643 hsm: Make sanity-hsm test_60 more robust
The first version of this test was fragile and could fail
intermittently when test infrastructure was not capable of
providing 1MB/sec in lustre bandwidth. This commit changes the
test to validate that a progress update occurs within the expected
window, rather than testing for a specific amount of data copied
under ideal conditions.

Lustre-commit: 1c9fbda98bf2274d63517ad72ef81484081834b9
Lustre-change: http://review.whamcloud.com/#/c/9376/

Signed-off-by: Michael MacDonald <michael.macdonald@intel.com>
Change-Id: I335e4cd9a3671f60be5a9236f322705b68664c8c
Reviewed-on: http://review.whamcloud.com/9378
Reviewed-by: Bob Glossman <bob.glossman@intel.com>
Tested-by: Jenkins
Reviewed-by: Jian Yu <jian.yu@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>