LU-4647 nodemap: add mapping functionality
Add functions to allow the MDT to use the nodemap module
functions to map UIDs and GIDs. When a client connects to the MDT,
the NID of the client is classified by nodemap and a pointer
to the nodemap is added to the export structure.
In old_init_ucred and old_init_ucred_reint, the body ids and fs
ids are mapped according to the nodemap state (enabled, maps,
squashed ids, trust_client_ids and allow_root_access). ACLs are mapped
in mdt_getattr_internal. Squashed ACLs are discarded before the ACLs
are sent to clients. Trusted administrative nodes are allowed to
use setfacl on all users. Mapped nodes can only use setfacl for users
also mapped into the node.

Change-Id: I9648f6837f8bb4ff145c8a0d647fc934c08aaffe
Signed-off-by: Joshua Walgenbach <jjw@iu.edu>
Signed-off-by: Kit Westneat <kit.westneat@gmail.com>
Reviewed-on: http://review.whamcloud.com/9299
Tested-by: Jenkins
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>
Reviewed-by: Nathaniel Clark <nathaniel.l.clark@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>