LU-4838 osd: object iterator in zfs to sync before iteration
The issue with dmu_object_next() is that it doesn't expose
non-committed blocks in the metadnode. Therefore the iterator
doesn't find dnodes allocated in the current (non-committed)
txg which breaks testing and may affect LFSCK. As a short-term
solution, osd_zfs_otable_it_init() ensures the current txg
is committed.

Signed-off-by: Alex Zhuravlev <alexey.zhuravlev@intel.com>
Change-Id: I639f37ced5790d1e1514c3e24594e360fcd6c1a8
Signed-off-by: Nathaniel Clark <nathaniel.l.clark@intel.com>
Reviewed-on: http://review.whamcloud.com/9838
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>
Tested-by: Andreas Dilger <andreas.dilger@intel.com>