LU-3953 build: fix paths for generated headers
Headers that are generated during kernel build and placed
in various include/generated directories are always in
the kernel build dir. There are several places in which the kernel
source dir is used instead. This causes configure/build failures
on systems using separate build and source directories.

This patch cleans those paths up.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>
Change-Id: I7335a8088d25f1e3bf64d8c131b542dfce38fff0
Reviewed-on: http://review.whamcloud.com/10160
Reviewed-by: Yang Sheng <yang.sheng@intel.com>
Reviewed-by: James Simmons <uja.ornl@gmail.com>
Reviewed-by: Dmitry Eremin <dmitry.eremin@intel.com>
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>