LU-4974 lod: add prefix to struct lod_pool_iterator
Add "lpi_" prefix to all fields in struct lod_pool_iterator (formerly
just "pool_iterator") so that the fields are easier to identify,
instead of having to grep for common names.

Fix up some CERROR() messages to be CDEBUG(), since they can be
printed under normal usage and are not very critical.

Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Change-Id: I5d271eec836622401a59d30edd045e220e500c1e
Reviewed-on: http://review.whamcloud.com/10245
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Dmitry Eremin <dmitry.eremin@intel.com>
Reviewed-by: John L. Hammond <john.hammond@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>