LU-5140 build: Make ko2iblnd buildable with Mellanox and newer kernels
The libcfs layer handles different level of support for process
namespace support. It does this by examining the uidgid.h
header. Both newer kernels and the Mellanox external OFED stack
supply their own uidgid.h header. In the current ko2iblnd
driver it completely ignores the Mellanox uidgid.h version
which ends up conflicting with the native uidgid.h header. The
reason libcfs misses the Mellanox handling is because the libcfs
and lnet headers are placed before the Mellanox headers so the
definations are missed. This patch moves all the libcfs and lnet
headers after the Mellanox header section.

Change-Id: Ibd7425b4a4d5b0f59c48e2007d870c67adb5f63d
Signed-off-by: James Simmons <uja.ornl@gmail.com>
Reviewed-on: http://review.whamcloud.com/10571
Tested-by: Jenkins
Reviewed-by: Bob Glossman <bob.glossman@intel.com>
Reviewed-by: Liang Zhen <liang.zhen@intel.com>
Reviewed-by: Shuichi Ihara <sihara@ddn.com>
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Patrick Farrell <paf@cray.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>