LUDOC-244 lfsck: Update lfsck documentation
Renamed 'Synopsis' sections to 'Usage' and swapped order with the
'Description' section. It makes more sense to see the description of a
feature before an example of its usage.

Renamed the 'Mount Options' section to 'Auto scrub' because it really
describes the auto_scrub parameter and what it does. Added a reference
in this section to the mount.lustre section where an explanation of
the noscrub mount option was added. The relationship between the
auto_scrub parameter and noscrub mount option is explained.

Also put 'N' inside a replaceable tag in the Auto scrub parameters
section.  This is meant to represent an integer.

Signed-off-by: Ryan Haasken <haasken@cray.com>
Change-Id: Ie947f8b268af299a0f583b98bb559f9d3313c0bf
Reviewed-on: http://review.whamcloud.com/11886
Tested-by: Jenkins
Reviewed-by: Richard Henwood <richard.henwood@intel.com>