LUDOC-247 mount: Clarify recovery_time_{soft,hard} defaults
It was not clear that the recovery_time_{soft,hard} default values
depended on the RPC timeout value at mount time.  The descriptions of
the default values of these recovery times were clarified.  Links were
added from the recovery_time_{soft,hard} descriptions to the section
which describes the RPC timeout.

Signed-off-by: Ryan Haasken <haasken@cray.com>
Change-Id: Iecf2814fb62e7bedff79873bbd802fc27277aafa
Reviewed-on: http://review.whamcloud.com/11885
Tested-by: Jenkins
Reviewed-by: Chris Horn <hornc@cray.com>
Reviewed-by: Richard Henwood <richard.henwood@intel.com>