LU-2456 lnet: configure lnet on startup
This is the ninth patch of a set of patches that enables DLC.

Modified lustre/scripts/lnet to look for a default config file
/etc/sysconfig/lnet.conf.  If it's found and the lnetctl utility
is installed, then call lnetctl with that file as a parameter.

This file is expected to be in YAML format and it defines
LNET configuration items.

Signed-off-by: Amir Shehata <amir.shehata@intel.com>
Change-Id: I32252abfddeaefd43359c7055a5476b2bc8eda92
Reviewed-on: http://review.whamcloud.com/11798
Tested-by: Jenkins
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>
Reviewed-by: James Simmons <uja.ornl@gmail.com>
Tested-by: Andreas Dilger <andreas.dilger@intel.com>