LU-5474 tests: sanity-hsm test_90 use local HSM_ARCHIVE
sanity-hsm test 90 suffers from frequent failures due to
slow archive speeds. If the existing archive is not
local, test 90 now uses a local disk archive to speed
the archive process.

sanity-hsm test 40 was modified to query the SINGLEAGT
node to check if the archive is a local disk for
HSM_ARCHIVE.

copytool_cleanup was modified to match copytool_setup;
remove the contents of $hsm_root and not $hsm_root itself.

This patch is back-ported from the following one:
Lustre-commit: 8641822bdb248a034bf4065c7e2e60cdb7a47041
Lustre-change: http://review.whamcloud.com/12069

Signed-off-by: James Nunez <james.a.nunez@intel.com>
Change-Id: I3aeaedf21d4b6a1316a9510f8e6f3d19662ccdec
Reviewed-on: http://review.whamcloud.com/12963
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Jinshan Xiong <jinshan.xiong@intel.com>
Reviewed-by: Faccini Bruno <bruno.faccini@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>