LUDOC-260 monitoring: clarify enable/disable jobstats
State where lctl commands must be run, to enable or disable collection
of jobstats.  Improve the description of JobID usage, and describe the
possibility of using separate jobid_var on different nodes.

Update jobstat examples to show output from multiple targets.

Fix a defect in the example for how to clear a single job stat.

Signed-off-by: Olaf Faaland <faaland1@llnl.gov>
Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Change-Id: I1866b1bff291fd3741ee751d60c7a5aae56c6084
Reviewed-on: http://review.whamcloud.com/13274
Tested-by: Jenkins
Reviewed-by: Niu Yawei <yawei.niu@intel.com>