LU-2524 test: Clean up replay-ost-single test suite
Miscellaneous cleanup of replay-ost-single including:
Replace `lfs setstripe` with $SETSTRIPE
Replace `lfs getstripe` with $GETSTRIPE
Replaced beginning of line space indent with tabs
Replaced ‘return’ with a call to ‘error’
Added call to `error` with error messages for a variety of common
routines ,like mkdir, or for functions that return a value.
Replace `…` with $(...)
Removed linefeed escape after |, ||, & and && operators.

Signed-off-by: James Nunez <james.a.nunez@intel.com>
Change-Id: I92bebb5db0b42d8e0fd23aa6c783d546a8fd6507
Reviewed-on: http://review.whamcloud.com/14685
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Jian Yu <jian.yu@intel.com>
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>