LU-6356 tests: fix sanity-krb5
test_151 is updated because it was relying on the fact that a target
cannot start if it cannot reaches MGS, which is no longer the case.
Now start the target, and look at import state of the mgc connection,
thanks to the newly added wait_mgc_import_state() routine in the
test framework.

Signed-off-by: Sebastien Buisson <sebastien.buisson@bull.net>
Change-Id: I8a846f1dfefe5d1d2310295bdd5172feb04739ef
Reviewed-on: http://review.whamcloud.com/15556
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Dmitry Eremin <dmitry.eremin@intel.com>
Reviewed-by: John L. Hammond <john.hammond@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>