LU-6895 osd-ldiskfs: handle REMOTE_PARENT_DIR properly
For a remote directory D_n, its name entry resides on the MDT_x,
its object resides on the MDT_y. On the MDT_y, the D_n's object
is inserted under REMOTE_PARENT_DIR and its dotdot entry is set
as the REMOTE_PARENT_DIR to satisfy local e2fsck.

When the namespace LFSCK scans the D_n directory on the MDT_y,
the REMOTE_PARENT_DIR will be found with LMAC_NOT_IN_OI in LMA,
the osd_dirent_check_repair() should NOT assert for such case.

Signed-off-by: Fan Yong <fan.yong@intel.com>
Change-Id: Id7fc10caae81be7d751c338fda401a6be93562ef
Reviewed-on: http://review.whamcloud.com/16035
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>
Reviewed-by: wangdi <di.wang@intel.com>
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>