LU-7109 lfsck: update OST-index in IDIF inside OSD
Old IDIF used "0" as the OST index, that may cause compatibility
issues when update. There is switch inside osd-ldiskfs for controlling
whether convert old IDIF to new one. Once the real OST index became
part of the IDIF and stored on disk (in LMA EA), then the OST cannot
be downgraded.

Because the conversion switch is inside osd-ldiskfs, the LFSCK should
not update the IDIF-in-LMA, instead, leave it to be handled by OSD to
avoid downgrading trouble.

Signed-off-by: Fan Yong <fan.yong@intel.com>
Change-Id: I9fee60ffabf732c0ab0734b1184e4e49638c9e88
Reviewed-on: http://review.whamcloud.com/16282
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Alex Zhuravlev <alexey.zhuravlev@intel.com>
Reviewed-by: Niu Yawei <yawei.niu@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>