LU-7962 build: Support builds w/ weak module ZFS
Now that ZFS supports weak module updates, ZFS does will not
always be built for every specific kernel that is installed.
Unfortunately, some directories that Lustre needs from ZFS and
SPL have a kernel string embedded in them, and now that kernel
string will potentially be different than the kernel that
Lustre is building against.

As a work around we can update the Lustre build scripts to just
select the last subdirectory (after all other methods fail) that
has a kernel name from the ZFS and SPL build directories as a
directory that Lustre should use in its build. Its as reasonable
a guess as I can figure we can easily make.

Change-Id: Id17940b78b99ddb8b589d1f46f85c3282c71a3ad
Signed-off-by: Christopher J. Morrone <morrone2@llnl.gov>
Reviewed-on: http://review.whamcloud.com/19249
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Nathaniel Clark <nathaniel.l.clark@intel.com>
Reviewed-by: Faccini Bruno <bruno.faccini@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>