LU-7060 ldev: Added MGS NID substitution to ldev
For automation purposes, it would be nice to have a
command substitution for the MGS NID of a node.
The ldev perl script now has a %m flag to sub in
the MGS NID. The MGS associated with a node is
determined by which file system the node is a part
of. The NID is obtained by using the current
NID file parsing. Also, MGS is now an acceptable
type to have specified in a label.

Signed-off-by: Giuseppe Di Natale <dinatale2@llnl.gov>
Change-Id: Ifcfad7664ac028e2074030e4b3ba5995ddb5358d
Reviewed-on: http://review.whamcloud.com/16129
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>
Reviewed-by: Bob Glossman <bob.glossman@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>