LU-3289 gss: Interface and code changes for shared key
This patch moves and modifies some of the kerberos code to
be more general so it can be utilized by the sk and null
gss implementations.  This includes the userspace portion
for lsvcgssd and the kernel pipefs interface.  Since the gss
interface has changed the version has also been incremented.

Signed-off-by: Jeremy Filizetti <jeremy.filizetti@gmail.com>
Change-Id: Ib46b74c9b0e01c4f71f16d288eae83c5c1567bfb
Reviewed-on: http://review.whamcloud.com/16728
Tested-by: Jenkins
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>
Tested-by: Andreas Dilger <andreas.dilger@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>