LU-7342 test: test_85a: unused locks error
When we changed FSNAME from lustre to lustre-1, $3
will give “addr” value “clilov” instead of getting
proper address, due to this we are getting "Found no match" error.
But if we change $3 to $NF getting addr value correctly.
Generally NF will give total number of fields in the current
input, but in our case it happens to be last field.

Test-Parameters: trivial testlist=replay-single

Seagate-bug-id: MRP-2791
Signed-off-by: Chennaiah Palla <chennaiah.palla@seagate.com>
Change-Id: I1d5e95425fdf848907cb9617494aa1c7971d90a6
Reviewed-on: http://review.whamcloud.com/16953
Reviewed-by: Jian Yu <jian.yu@intel.com>
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Saurabh Tandan <saurabh.tandan@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>