LU-3289 gss: Add userspace support for GSS null and sk
Add the support to lgss_keyring (client) and lsvcgssd (server) for GSS
null and shared key.  This includes the common functionality for the
creation of the session key in privacy mode and shared key for
integrity and privacy mode.

A new utility lgss_sk has been added for key management.

Signed-off-by: Jeremy Filizetti <jeremy.filizetti@gmail.com>
Change-Id: Id2574b940d344d18f793113f61e69a393b6df4f0
Reviewed-on: http://review.whamcloud.com/17600
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>
Tested-by: Jenkins
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>