LU-3289 gss: Add option for loading keys during mount
This adds functionality for loading GSS shared keys using
mount.lustre (mount) command.  The --skpath option has been
added which supports a directory or file of keys to load.
The context from which the mount occurs is used to determine
which type of key to load (MGS, MDT/OST, or client).

Signed-off-by: Jeremy Filizetti <jeremy.filizetti@gmail.com>
Change-Id: I4ffbfb9ea76d32ed69f94da93f82e90e161cb99d
Reviewed-on: http://review.whamcloud.com/17601
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>
Tested-by: Jenkins
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>