LU-7591 test: Applying filtering in t32_test
With SELinux feature enabled on Client, conf-sanity
test_32b failed with 'list verification failed' and
'Host key verification failed'. Because of the SELinux
enabled feature there was extra '.' in the permission
column.

Hence, that filtering is applied to remove that extra
'.' from permission column. t32_test within conf-sanity
is modified by applying filter to it.

Signed-off-by: Saurabh Tandan <saurabh.tandan@intel.com>
Change-Id: I62fbb79ecf29bb09fd382ab1ebc62d41f6678280
Reviewed-on: http://review.whamcloud.com/17973
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Jian Yu <jian.yu@intel.com>
Reviewed-by: James Nunez <james.a.nunez@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>