LU-7735 utils: Allow revert of NO_PRIMNODE
When formatting a target or modifying an existing one with the
parameter "--servicenode", the flag no_primnode is set on the target,
modifying the way the target is then registered on the MGS to avoid
adding the NID of the node actually registering and let only the nids
given with --servicenode parameters.

However, if someone wants to revert to the older behavior with the
"--failnode" parameter, the no_primnode flag is never unset, and the
primary node NID is never added to the mountdata.

Test-Parameters: trivial testlist=conf-sanity
Signed-off-by: Nathaniel Clark <nathaniel.l.clark@intel.com>
Change-Id: I2251f66236c91365c03e7d138cd8e6508956d73d
Reviewed-on: https://review.whamcloud.com/18304
Reviewed-by: Bob Glossman <bob.glossman@intel.com>
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>