LU-7756 oss: allow larger number of OSS service threads
The current static limit on the maximum number of OSS service threads
is sometimes not large enough to have enough requests sent to the
OST devices.

This patch makes this limit tunable with a new module parameter
oss_max_threads, so that it can be adapted to the underlying
storage hardware.

Signed-off-by: Gregoire Pichon <gregoire.pichon@bull.net>
Change-Id: I182d1dc0a35b42b85b9c0e2a5413e3a852a4e5c3
Reviewed-on: http://review.whamcloud.com/18350
Tested-by: Jenkins
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>
Reviewed-by: Bob Glossman <bob.glossman@intel.com>
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>