LU-5092 nodemap: transfer nodemaps between MGS, MDTs, and OSTs
This creates a new config lock on the MGS for transferring the
nodemaps.  The target MGCs can enqueue this lock, which is then
revoked by the MGS when the nodemaps are modified. The MGCs then
issue a read config RPC on the nodemap config lock, which causes
the MGS to transfer the nodemap index file to the MGCs. The MGCs
then process the index file, similarly to the way the MGS
processes it on start.

Signed-off-by: Kit Westneat <kit.westneat@gmail.com>
Change-Id: I99a34773298484f70a912761f5831b75196d41d8
Reviewed-on: http://review.whamcloud.com/11830
Reviewed-by: John L. Hammond <john.hammond@intel.com>
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: James Simmons <uja.ornl@yahoo.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>