LU-7890 lov: Ensure correct operation for large object sizes
If a backing filesystem (ZFS) returns that it supports very large
(LLONG_MAX) object sizes, that should be correctly supported.  This
fixes the check for unitialized stripe_maxbytes in
lsm_unpackmd_common(), so that ZFS can return LLONG_MAX and it will be
okay.  This issue is excersized by writing to or past the 2TB boundry
of a singly stripped file.  A regression test is added to check for
this.

Signed-off-by: Nathaniel Clark <nathaniel.l.clark@intel.com>
Change-Id: I669096281bd826d69b5e86ada50077740412d268
Reviewed-on: http://review.whamcloud.com/19066
Reviewed-by: Andreas Dilger <andreas.dilger@intel.com>
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>
Reviewed-by: Jinshan Xiong <jinshan.xiong@intel.com>
Reviewed-by: Oleg Drokin <oleg.drokin@intel.com>