LUDOC-306 dne: add more description for DNE usage
Add some more references to "DNE" in the text, instead of only in the
glossary, so that it is more easily found when searching the manual.

Add some more examples and explanation of how and when to use multiple
MDTs and striped vs. remote directories.  Definitely more could be
written on this topic, but it is a start.

Signed-off-by: Andreas Dilger <andreas.dilger@intel.com>
Change-Id: Iaec344afe2d909ece0a1935eb33d8a58ad992370
Reviewed-on: http://review.whamcloud.com/19177
Tested-by: Jenkins
Reviewed-by: Richard Henwood <richard.henwood@intel.com>